Revert "Feat/coins"
