+++ ios/Flutter/Debug.xcconfig
@@ -1,2 +1,1 @@
-#include? "Pods/Target Support Files/Pods-Runner/Pods-Runner.debug.xcconfig"
 #include "Generated.xcconfig"
