--- lib/core/styles/typography.dart
@@ -1,210 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
-
-class RecycleRushHeadingTypography {
-  const RecycleRushHeadingTypography(this.color);
-  final Color color;
-
-  get textColor => color;
-  static TextStyle megaLarge({Color? color}) => GoogleFonts.spaceGrotesk(
-        fontSize: 48,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: -0.02,
-        color: color,
-      );
-
-  static TextStyle mega({Color? color}) => GoogleFonts.spaceGrotesk(
-        fontSize: 48,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: -0.02,
-        color: color,
-      );
-
-  static TextStyle xxLarge({Color? color}) => GoogleFonts.spaceGrotesk(
-        fontSize: 40,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: -0.02,
-        color: color,
-      );
-
-  static TextStyle xLarge({Color? color}) => GoogleFonts.spaceGrotesk(
-        fontSize: 36,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: -0.02,
-        color: color,
-      );
-
-  static TextStyle large({Color? color}) => GoogleFonts.spaceGrotesk(
-        fontSize: 32,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: -0.02,
-        color: color,
-      );
-
-  static TextStyle medium({Color? color}) => GoogleFonts.spaceGrotesk(
-        fontSize: 28,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: -0.02,
-        color: color,
-      );
-
-  static TextStyle small({Color? color}) => GoogleFonts.spaceGrotesk(
-        fontSize: 24,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: -0.02,
-        color: color,
-      );
-
-  static TextStyle xSmall({Color? color}) => GoogleFonts.spaceGrotesk(
-        fontSize: 20,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: -0.02,
-        color: color,
-      );
-
-  static TextStyle xxSmall({Color? color}) => GoogleFonts.spaceGrotesk(
-        fontSize: 16,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: -0.02,
-        color: color,
-      );
-}
-
-class RecycleRushSubHeadingTypography {
-  const RecycleRushSubHeadingTypography(this.color);
-  final Color color;
-
-  get textColor => color;
-  static TextStyle xxLarge({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 24,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: 0,
-        color: color,
-      );
-
-  static TextStyle xLarge({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 20,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: 0,
-        color: color,
-      );
-
-  static TextStyle large({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 18,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: 0,
-        color: color,
-      );
-
-  static TextStyle medium({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 16,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: 0,
-        color: color,
-      );
-
-  static TextStyle small({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 14,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: 0,
-        color: color,
-      );
-
-  static TextStyle xSmall({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 12,
-        fontWeight: FontWeight.w700,
-        height: 1.2,
-        letterSpacing: 0,
-        color: color,
-      );
-}
-
-class RecycleRushLabelTypography {
-  const RecycleRushLabelTypography(this.color);
-  final Color color;
-
-  get textColor => color;
-  static TextStyle large({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 18,
-        fontWeight: FontWeight.w500,
-        height: 1.5,
-        letterSpacing: 0,
-        color: color,
-      );
-
-  static TextStyle medium({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 16,
-        fontWeight: FontWeight.w500,
-        height: 1.5,
-        letterSpacing: 0,
-        color: color,
-      );
-
-  static TextStyle small({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 14,
-        fontWeight: FontWeight.w500,
-        height: 1.5,
-        letterSpacing: 0,
-        color: color,
-      );
-
-  static TextStyle xSmall({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 12,
-        fontWeight: FontWeight.w500,
-        height: 1.5,
-        letterSpacing: 0,
-        color: color,
-      );
-}
-
-class RecycleRushParagraph {
-  const RecycleRushParagraph(this.color);
-  final Color color;
-
-  get textColor => color;
-  static TextStyle large({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 18,
-        fontWeight: FontWeight.w400,
-        height: 1.5,
-        letterSpacing: 0,
-        color: color,
-      );
-
-  static TextStyle medium({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 16,
-        fontWeight: FontWeight.w400,
-        height: 1.5,
-        letterSpacing: 0,
-        color: color,
-      );
-
-  static TextStyle small({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 14,
-        fontWeight: FontWeight.w400,
-        height: 1.5,
-        letterSpacing: 0,
-        color: color,
-      );
-
-  static TextStyle xSmall({Color? color}) => GoogleFonts.plusJakartaSans(
-        fontSize: 12,
-        fontWeight: FontWeight.w400,
-        height: 1.5,
-        letterSpacing: 0,
-        color: color,
-      );
-}
