--- lib/feature/coins/presentation/view/coins_view.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:recycle_rush_fe/feature/coins/presentation/widget/coins_widget.dart';
-
-class RecycleRushViewCoins extends StatelessWidget {
-  static const routeName = '/coins';
-  const RecycleRushViewCoins({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const RecycleRushWidgetCoins();
-  }
-}
