--- lib/feature/home/presentation/view/home_view.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:recycle_rush_fe/feature/home/presentation/widget/home_widget.dart';
-
-class RecycleRushViewHome extends StatelessWidget {
-  static const routeName = '/home';
-  const RecycleRushViewHome({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const RecycleRushWidgetHome();
-  }
-}
